Initial update for Julia 0.3.1.

+++ src/bloom-filter.jl
@@ -15,8 +15,8 @@ end
 # which uses 2 hash functions vs. k and has comparable properties
 # See Kirsch and Mitzenmacher, 2008: http://www.eecs.harvard.edu/~kirsch/pubs/bbbf/rsa.pdf
 function hash_n(key::String, k::Int, max::Int)
-    a_hash = hash(key, 0)
-    b_hash = hash(key, 170)
+    a_hash = hash(key, uint(0))
+    b_hash = hash(key, uint(170))
     hashes = Array(Uint, k)
     for i in 1:k
         hashes[i] = mod(a_hash + i * b_hash, max) + 1
